move user and auth views to the appropriate package
